requirement.txt
pandas
Pyarrow
numpy
virtualenv
scikit-learn
fastapi
uvicorn
python-dotenv
SQLAlchemy
seaborn
matplotlib
psycopg2
pickle4
plotly
flask
flask_sqlalchemy
requests
azure-cognitiveservices-speech
dateparser
spacy
geopy
datefinder
transformers
fuzzywuzzy
python-Levenshtein

